+++ esbuild.config.mjs
@@ -44,7 +44,6 @@ const context = await esbuild.context({
 		'process.env.INSTAPAPER_DEBUG': JSON.stringify(process.env['INSTAPAPER_DEBUG'] ?? !prod),
 		'process.env.INSTAPAPER_BASE_URL': JSON.stringify(process.env['INSTAPAPER_BASE_URL'] ?? ''),
 		'process.env.INSTAPAPER_CONSUMER_KEY': JSON.stringify(process.env['INSTAPAPER_CONSUMER_KEY']),
-		'process.env.INSTAPAPER_CONSUMER_SECRET': JSON.stringify(process.env['INSTAPAPER_CONSUMER_SECRET']),
 	},
 });